Version finale du project sans check de la moulinette

+++ srcs/bresenham.c
@@ -6,7 +6,7 @@
 /*   By: vnaoussi <vnaoussi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/03 23:50:47 by vnaoussi          #+#    #+#             */
-/*   Updated: 2026/02/06 09:56:23 by vnaoussi         ###   ########.fr       */
+/*   Updated: 2026/02/08 01:01:24 by vnaoussi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,19 +44,19 @@ void	draw_lign(t_img *img, t_proj_dot dot_a, t_proj_dot dot_b)
 	int		delta[2];
 	int		i;
 
-	step[0] = get_step((int)dot_a.proj_x, (int)dot_b.proj_x);
-	step[1] = get_step((int)dot_a.proj_y, (int)dot_b.proj_y);
-	delta[0] = abs((int)dot_b.proj_x - (int)dot_a.proj_x);
-	delta[1] = abs((int)dot_b.proj_y - (int)dot_a.proj_y);
+	step[0] = get_step(dot_a.proj_x, dot_b.proj_x);
+	step[1] = get_step(dot_a.proj_y, dot_b.proj_y);
+	delta[0] = abs(dot_b.proj_x - dot_a.proj_x);
+	delta[1] = abs(dot_b.proj_y - dot_a.proj_y);
 	err[0] = delta[0] - delta[1];
 	i = 0;
 	while (1)
 	{
 		percent = (double)i++ / fmax(delta[0], delta[1]);
-		put_pixel(img, (int)dot_a.proj_x, (int)dot_a.proj_y,
-				get_color((int)dot_a.color, (int)dot_b.color, percent));
-		if ((int)dot_a.proj_x == (int)dot_b.proj_x
-				&& (int)dot_a.proj_y == (int)dot_b.proj_y)
+		put_pixel(img, dot_a.proj_x, dot_a.proj_y,
+				get_color(dot_a.color, dot_b.color, percent));
+		if (dot_a.proj_x == dot_b.proj_x
+				&& dot_a.proj_y == dot_b.proj_y)
 			return ;
 		err[1] = 2 * err[0];
 		if (err[1] > -delta[1])
